REST-243: Add new name for project and delete unnecessary files
