+++ .gitignore
@@ -33,7 +33,7 @@ bld/
 [Ll]ogs/
 
 # Visual Studio 2015/2017 cache/options directory
-DotnetBoilerPlate/.vs
+DotnetFoundation/.vs
 # Uncomment if you have tasks that create the project's static files in wwwroot
 #wwwroot/
