--- DotnetBoilerPlate/DotnetBoilerPlate.API/Helpers/AutoMappingProfile.cs
@@ -1,8 +1,0 @@
-using AutoMapper;
-
-namespace DotnetBoilerPlate.API.Helpers
-{
-    public class AutoMappingProfile : Profile
-    {
-    }
-}
