--- DotnetBoilerPlate/DotnetBoilerPlate.API/Helpers/EnumHelper.cs
@@ -1,15 +1,0 @@
-namespace DotnetBoilerPlate.API.Helpers
-{
-    public static class EnumHelper
-    {
-        public static string ConvertToString(Enum enumElement)
-        {
-            return Enum.GetName(enumElement.GetType(), enumElement);
-        }
-
-        public static TEnumType ConvertToEnum<TEnumType>(string enumValue)
-        {
-            return (TEnumType)Enum.Parse(typeof(TEnumType), enumValue);
-        }
-    }
-}
