--- DotnetBoilerPlate/DotnetBoilerPlate.DAL/Models/TestUserRegistrationdetailsDBO.cs
@@ -1,10 +1,0 @@
-namespace DotnetBoilerPlate.DAL.Models;
-
-public partial class TestUserRegistrationdetailsDBO
-{
-    public string? HtNo { get; set; }
-
-    public string? FailedDbReason { get; set; }
-
-    public string? InsertStmt { get; set; }
-}
