--- DotnetBoilerPlate/DotnetBoilerPlate.DAL/Models/TrainingprovidersDBO.cs
@@ -1,19 +1,0 @@
-namespace DotnetBoilerPlate.DAL.Models;
-
-public partial class TrainingprovidersDBO
-{
-    public string TrainingProviderId { get; set; } = null!;
-
-    /// <summary>
-    /// 1 - active, 2 - deactive, 3 - deleted
-    /// </summary>
-    public bool? Status { get; set; }
-
-    public string? CreatedBy { get; set; }
-
-    public DateTime? CreatedOn { get; set; }
-
-    public string? ModifiedBy { get; set; }
-
-    public DateTime? ModifiedOn { get; set; }
-}
