--- DotnetBoilerPlate/DotnetBoilerPlate.DAL/Repositories/Interfaces/IUsersRepo.cs
@@ -1,10 +1,0 @@
-using DotnetBoilerPlate.DAL.Models;
-
-namespace DotnetBoilerPlate.DAL.Repositories.Interfaces
-{
-    public interface IUsersRepo
-    {
-        List<UsersDBO> GetAllUsers();
-        UsersDBO GetUserById(string userId);
-    }
-}
