+++ DotnetFoundation/.editorconfig
