+++ DotnetFoundation/DotnetFoundation.API/BLL/Interfaces/IUsersBLL.cs
@@ -1,0 +1,12 @@
+using Microsoft.AspNetCore.Mvc;
+using DotnetFoundation.API.Models;
+using DotnetFoundation.DAL.Models;
+
+namespace DotnetFoundation.API.BLL.Interfaces
+{
+    public interface IUsersBLL
+    {
+        ActionResult<BaseResponse<List<UsersDBO>>> GetAllUsers();
+        ActionResult<BaseResponse<UsersDBO>> GetUserById(string userId);
+    }
+}
