+++ DotnetFoundation/DotnetFoundation.API/Helpers/AutoMappingProfile.cs
@@ -1,0 +1,8 @@
+using AutoMapper;
+
+namespace DotnetFoundation.API.Helpers
+{
+    public class AutoMappingProfile : Profile
+    {
+    }
+}
