+++ DotnetFoundation/DotnetFoundation.API/Helpers/EnumHelper.cs
@@ -1,0 +1,15 @@
+namespace DotnetFoundation.API.Helpers
+{
+    public static class EnumHelper
+    {
+        public static string ConvertToString(Enum enumElement)
+        {
+            return Enum.GetName(enumElement.GetType(), enumElement);
+        }
+
+        public static TEnumType ConvertToEnum<TEnumType>(string enumValue)
+        {
+            return (TEnumType)Enum.Parse(typeof(TEnumType), enumValue);
+        }
+    }
+}
