+++ DotnetFoundation/DotnetFoundation.DAL/Repositories/Interfaces/IUsersRepo.cs
@@ -1,0 +1,10 @@
+using DotnetFoundation.DAL.Models;
+
+namespace DotnetFoundation.DAL.Repositories.Interfaces
+{
+    public interface IUsersRepo
+    {
+        List<UsersDBO> GetAllUsers();
+        UsersDBO GetUserById(string userId);
+    }
+}
