+++ DotnetFoundation/DotnetFoundation.sln
@@ -1,0 +1,36 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.7.34003.232
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DotnetFoundation.API", "DotnetFoundation.API\DotnetFoundation.API.csproj", "{1C10A534-DB66-494B-B797-55143C40C9D1}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DotnetFoundation.DAL", "DotnetFoundation.DAL\DotnetFoundation.DAL.csproj", "{9E92CC0E-F0F8-4511-A5D9-D7B192CC9DA1}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{37C6C65D-EB24-40A7-A954-0446B3D1511D}"
+	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{1C10A534-DB66-494B-B797-55143C40C9D1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1C10A534-DB66-494B-B797-55143C40C9D1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1C10A534-DB66-494B-B797-55143C40C9D1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1C10A534-DB66-494B-B797-55143C40C9D1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9E92CC0E-F0F8-4511-A5D9-D7B192CC9DA1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9E92CC0E-F0F8-4511-A5D9-D7B192CC9DA1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9E92CC0E-F0F8-4511-A5D9-D7B192CC9DA1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9E92CC0E-F0F8-4511-A5D9-D7B192CC9DA1}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {52544ABA-43AE-4042-B5BF-D7287DD51EDD}
+	EndGlobalSection
+EndGlobal
